Reorganize components for TabColumnPicking
